requirements.txt
aiohttp>=3.9.0
websockets>=12.0
python-telegram-bot>=20.7
numpy>=1.26.0
pandas>=2.1.0
ccxt>=4.2.0
python-dotenv>=1.0.0
